Code Update Issue #11 and Solve Code-Effeciency The Issue #11 Output Repeat Bug is Resloved. Removed All In-Function Self-Calls in Player(), and replaced them with do-while Loop. Its not Fixed on some Functions (on Purpose), because it makes the UI a little Better in those Functions.
The Efficiency PROBLEMS Solved:
So Now, the Range of Players can be easily Changed, but for Now, it's set to 8.
    The Default Player Symbol Marks are Assigned in the Initialization.
    There's No Need to Assign a Player-Number everyTime. It just Needs a Check to reset the Player-Number to 1.

Quite a Few Other Changes.
The Docs() is Removed. The Play() is renamed Game(). The new PlayAgain() is defined.
The Custom Range of Konnect-Goal is Decreased to 12.
The Readme File is also Updated Accordingly.